+++ R/knnflex.R
@@ -7,9 +7,8 @@ classprob <- function(x) {
   x <- as.factor(x)
   n <- nlevels(x)
   votes <- rep(0, n)
-  for (i in 1:length(x)) votes[as.integer(x[i])] <- votes[as.integer(x[i])] +
-    1
-  votes/length(x)
+  for (i in 1:length(x)) votes[as.integer(x[i])] <- votes[as.integer(x[i])] + 1
+  votes / length(x)
 }
 
 #' @references https://github.com/cran/knnflex/blob/master/R/knn.dist.R