+++ R/splendid_ensemble.R
@@ -6,10 +6,10 @@
 #' @examples
 #' dat <- iris[, 1:4]
 #' class <- iris$Species
-#' sm <- splendid_model(dat, class, n = 3, algorithms = c("xgboost", "lda"))
+#' sm <- splendid_model(dat, class, n = 3, algorithms = c("xgboost", "slda"))
 #' se <- splendid_ensemble(sm, dat, class)
-splendid_ensemble <- function(sm, data, class, top = 3, seed = 1, rfe = FALSE,
-                              sequential = FALSE) {
+splendid_ensemble <- function(sm, data, class, top = 3, seed_rank = 1,
+                              rfe = FALSE, sequential = FALSE) {
   # vector of best performing algorithms from each bootstrap replicate
   bests <- sm$evals %>%
     do.call(cbind, .) %>%
@@ -18,13 +18,13 @@ splendid_ensemble <- function(sm, data, class, top = 3, seed = 1, rfe = FALSE,
     dplyr::mutate(logloss = -logloss,
                   boot = gsub(".*\\.", "\\1", rownames(.))) %>%
     split(.$boot) %>%
-    purrr::map(~ dplyr::select(.x, -dplyr::matches("\\.|boot"))) %>%
+    purrr::map(~ dplyr::select(., -dplyr::matches("\\.|boot"))) %>%
     purrr::map(~ purrr::map(.x, ~ names(sm$models)[order(
       rank(-.x, ties.method = "random"))])) %>%
-    purrr::map(~ purrr::invoke(rbind, .x)) %>%
+    purrr::map(~ purrr::invoke(rbind, .)) %>%
     purrr::map_chr(~ {
       if (ncol(.x) > 1) {
-        RankAggreg::RankAggreg(.x, ncol(.x), method = "GA", seed = seed,
+        RankAggreg::RankAggreg(.x, ncol(.x), method = "GA", seed = seed_rank,
                                verbose = FALSE)$top.list[1]
       } else {
         .x[1]
@@ -47,5 +47,5 @@ splendid_ensemble <- function(sm, data, class, top = 3, seed = 1, rfe = FALSE,
   seq_preds <- sequential %>%
     purrr::when(. ~ sequential_pred(seq_mods, sm, data, class))
 
-  dplyr::lst(bests, ensemble_algs, ensemble_mods, seq_mods, seq_preds)
+  tibble::lst(bests, ensemble_algs, ensemble_mods, seq_mods, seq_preds)
 }