+++ NEWS.md
@@ -1,3 +1,7 @@
 # splendid 0.1.0
 
+* Improved RFE interface
+
+* Added AdaBoost.M1 algorithm
+
 * Added a `NEWS.md` file to track changes to the package.